# openapi.yaml
openapi: 3.0.3
info:
  title: eBadge ID API
  description: >
    API pública del backend principal de eBadge ID. Cubre la superficie
    completa expuesta bajo `/api` por el backend principal: autenticación
    (local y SSO/SAML), organizaciones, usuarios, credenciales, plantillas
    de diseño, contratos digitales, metas/tareas/puntajes, notificaciones,
    API keys, invitaciones, activos de organización, brand kit, uploads,
    integraciones LMS, anclaje blockchain y los tres tableros de
    analítica/desempeño.

    Cerrado como parte del hallazgo H-15 de la auditoría técnica integral
    (2026-08): antes solo documentaba el subconjunto pensado para
    integración externa (credenciales/plantillas/auth). Cada endpoint de
    abajo se extrajo leyendo directamente su archivo de rutas y su
    controlador real — método, parámetros, autenticación, cuerpo de
    petición y códigos de respuesta reflejan el código tal como está, no
    un diseño aspiracional. Las mesa de ayuda (`help_backend`) es un
    servicio separado, con su propia base URL, y no está cubierta aquí.
  version: "1.1"
servers:
  - url: https://api.ebadgeid.com/api
    description: Producción
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token de sesión de usuario (POST /auth/login), enviado como
        `Authorization: Bearer <token>` o, alternativamente, como la cookie
        httpOnly `ebadge_token` que el propio login también setea.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Clave de integración por organización (ver POST /keys/{organization_code}).
        Alternativa a bearerAuth para llamadas programáticas — no requiere
        una sesión de usuario humano. Sujeta a su propio límite de
        requests/minuto (requests_allowed_per_minute).
  parameters:
    limitParam:
      name: limit
      in: query
      schema: { type: integer, default: 200, maximum: 500 }
      description: Máximo de resultados. Se recorta a 500 aunque se pida más.
    skipParam:
      name: skip
      in: query
      schema: { type: integer, default: 0 }
  schemas:
...
    Credential:
      type: object
      properties:
        credential_code:
          type: string
          example: CRED-A1B2C3D4E5F6
        credential_title:
          type: string
        credential_issue_date:
          type: string
          format: date
        credential_expiry_date:
          type: string
          format: date
        credential_pic_url:
          type: string
          format: uri
        credential_blockchain_hashes:
          type: string
          description: >
            Hash de integridad SHA-256, reproducible a partir de los campos
            persistidos del propio registro — no es un registro en una
            blockchain (ver AUDIT_FIXES.md). Se recalcula y compara en cada
            verificación.
        achiever_username:
          type: string
        organization_code:
          type: string
        credential_status:
          type: string
          enum: [Issued, Claimed, Expired, Revoked]
        integrity_valid:
          type: boolean
          description: >
            Solo presente en la respuesta de verificación pública. `false`
            significa que el hash recalculado no coincide con el guardado —
            el registro pudo haber sido alterado después de emitido.
security:
  - bearerAuth: []
paths:
...
  /credentials/by-code/{credential_code}:
    get:
      tags: [Credentials]
      summary: Verificar una credencial (público — sin autenticación)
      description: >
        El endpoint central del producto. Cualquiera con el código puede
        confirmar si una credencial es genuina. Recalcula el hash de
        integridad en cada llamada y lo compara contra el guardado —
        `integrity_valid: false` significa alteración posterior a la
        emisión.
      security: []
      parameters:
        - name: credential_code
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Credential' }
        '404': { description: Credencial no encontrada }
  /credentials/by-code/{credential_code}/openbadge:
    get:
      tags: [Credentials]
      summary: Exportar la misma credencial pública en formato Open Badges 3.0
      security: []
      parameters: [{ name: credential_code, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: "Documento Open Badges 3.0 (JSON-LD)" }
        '404': { description: Credencial no encontrada }
  /credentials/reserve-code:
    post:
      summary: Reservar un credential_code único antes de generar el certificado
      tags: [Credentials]
      description: >
        Paso 1 del flujo de emisión — reserva el código real antes de generar
        la imagen del certificado, para que el QR pueda apuntar al lugar
        correcto desde el principio. Ver AUDIT_FIXES.md, "Diseño de
        certificados".
      security: [{ bearerAuth: [] }, { apiKeyAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  credential_code: { type: string }
        '401': { description: No autenticado }
        '402': { description: Límite mensual de credenciales del plan alcanzado }
  /credentials/generate-certificate:
    post:
      tags: [Credentials]
      summary: Generar la imagen del certificado a partir de una plantilla
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [design_code, achiever_username, credential_code]
              properties:
                design_code: { type: string }
                achiever_username: { type: string }
                credential_code: { type: string, description: "Debe haberse reservado antes con POST /credentials/reserve-code." }
                guest_recipient: { type: object }
                custom_fields: { type: object }
      responses:
        '201': { description: "Imagen generada", content: { application/json: { schema: { type: object, properties: { url: { type: string, format: uri } } } } } }
        '400': { description: Campos faltantes o custom_fields inválidos }
        '404': { description: Plantilla o destinatario no encontrado }
        '409': { description: "La plantilla no está publicada" }
        '429': { description: "Límite de 60 certificados/15min excedido" }
        '502': { description: Falla el microservicio generador de certificados }
        '503': { description: "Generador o PUBLIC_APP_URL no configurados" }
  /credentials/create:
    post:
      summary: Emitir una credencial
      tags: [Credentials]
      description: Paso 2 del flujo — crea la credencial usando el código reservado en el paso 1. Acepta `achiever_username` (usuario real de la organización) o `guest_recipient` (nombre + email, sin cuenta).
      security: [{ bearerAuth: [] }, { apiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential_pic_url, achiever_username]
              properties:
                credential_code: { type: string, description: "El reservado en /credentials/reserve-code. Si se omite, se genera uno nuevo." }
                credential_pic_url: { type: string, format: uri, description: "Debe resolver a un origen de storage aprobado." }
                achiever_username: { type: string }
                credential_title: { type: string }
                guest_recipient:
                  type: object
                  properties:
                    first_name: { type: string, maxLength: 100 }
                    last_name: { type: string, maxLength: 100 }
                    email: { type: string, format: email }
                    designation: { type: string, maxLength: 100 }
                    city: { type: string, maxLength: 100 }
                custom_fields: { type: object }
      responses:
        '201':
          description: Credencial creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Credential' }
        '400': { description: "Faltan campos requeridos, URL de imagen no aprobada, o custom_fields inválidos" }
        '404': { description: Usuario u organización no encontrados }
        '409': { description: "credential_code ya en uso, o achiever_username de un guest_recipient coincide con una cuenta real" }
        '402': { description: Límite mensual de credenciales del plan alcanzado }
  /credentials/bulk-issue:
    post:
      tags: [Credentials]
      summary: Encolar la emisión masiva de credenciales
      description: >
        Encola el lote en BullMQ/Redis para procesamiento en segundo plano.
        Si la organización tiene `require_bulk_approval` activado, en vez de
        encolar de inmediato crea una solicitud pendiente (ver los 3
        endpoints de abajo).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipients:
                  type: array
                  maxItems: 2000
                  items: { type: object, properties: { achiever_username: { type: string }, guest_recipient: { type: object } } }
                usernames: { type: array, items: { type: string }, description: "Forma legacy, equivalente a recipients sin guest_recipient." }
                design_code: { type: string }
                credential_title: { type: string }
      responses:
        '202': { description: "Encolado (o `pending_approval` si la organización requiere aprobación)" }
        '400': { description: "recipients vacío, más de 2000 destinatarios, o falta achiever_username en alguno" }
        '403': { description: Requiere rol admin }
        '503': { description: "Cola no configurada (REDIS_URL no seteado)" }
  /credentials/bulk-issue/pending:
    get:
      tags: [Credentials]
      summary: Listar solicitudes de emisión masiva pendientes de aprobación
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /credentials/bulk-issue/{pendingId}/approve:
    post:
      tags: [Credentials]
      summary: Aprobar una emisión masiva pendiente (esto es lo que realmente la encola)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: pendingId, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Aprobada y encolada }
        '404': { description: Solicitud pendiente no encontrada }
        '503': { description: Cola no configurada }
  /credentials/bulk-issue/{pendingId}/reject:
    post:
      tags: [Credentials]
      summary: Rechazar una emisión masiva pendiente (queda como registro, nunca se encola)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: pendingId, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { reason: { type: string, maxLength: 500 } } } } }
      responses:
        '200': { description: Rechazada }
        '404': { description: Solicitud pendiente no encontrada }
  /credentials/bulk-issue/{batchId}:
    get:
      tags: [Credentials]
      summary: Estado de progreso de un lote de emisión masiva
      security: [{ bearerAuth: [] }]
      parameters: [{ name: batchId, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
        '404': { description: "Lote no encontrado (incluye lotes de otra organización)" }
        '503': { description: Cola no configurada }
  /credentials/by-organization/{organization_code}:
    get:
      tags: [Credentials]
      summary: Listar credenciales de una organización (paginado)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: organization_code, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/limitParam' }
        - { $ref: '#/components/parameters/skipParam' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Credential' } }
                  pagination: { type: object, properties: { total_items: { type: integer } } }
        '403': { description: De otra organización }
  /credentials/by-user/{username}:
    get:
      tags: [Credentials]
      summary: Credenciales de un usuario (alias también en /by-users/{username})
      security: [{ bearerAuth: [] }]
      parameters: [{ name: username, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
        '403': { description: "De otra organización, o de otro usuario sin ser admin" }
        '404': { description: "Usuario no encontrado, o sin credenciales" }
  /credentials/revoke/{credential_code}:
    put:
      summary: Revocar una credencial
      tags: [Credentials]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: credential_code
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Revocada }
        '400': { description: Ya estaba revocada }
        '404': { description: No encontrada }
  /credentials/{credential_code}/history:
    get:
      summary: Historial de auditoría de una credencial (emitida/reclamada/revocada, quién y cuándo)
      tags: [Credentials]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: credential_code
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  credential_code: { type: string }
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        event: { type: string, enum: [issued, claimed, revoked] }
                        actor_username: { type: string, nullable: true }
                        createdAt: { type: string, format: date-time }
        '403': { description: De otra organización }
        '404': { description: No encontrada }
  /credentials/claim/{credential_code}:
    put:
      tags: [Credentials]
      summary: Reclamar una credencial propia (Issued -> Claimed)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: credential_code, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Reclamada }
        '400': { description: "Ya estaba reclamada, revocada, o expiró" }
        '404': { description: No encontrada }
  /credentials/portal-visibility/{credential_code}:
    put:
      tags: [Credentials]
      summary: Mostrar/ocultar una credencial propia en el portal público del recipiente
      security: [{ bearerAuth: [] }]
      parameters: [{ name: credential_code, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [visible], properties: { visible: { type: boolean } } } } }
      responses:
        '200': { description: OK }
        '400': { description: "Falta el campo visible, o la credencial aún no está Claimed" }
        '404': { description: No encontrada }
  /credentials/analytics/organization/{organization_code}:
    get:
      tags: [Credentials]
      summary: Analítica de credenciales de la propia organización (emisor)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: organization_code, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
        '403': { description: De otra organización }
  /credentials/portal/{username}:
    get:
      tags: [Credentials]
      summary: Portal público de un recipiente (solo credenciales Claimed y visibles)
      security: []
      parameters: [{ name: username, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
        '404': { description: Sin credenciales públicas para este perfil }
        '429': { description: "Límite de 120 requests/15min excedido" }

  # ============================== DESIGNS ==============================
...